# openapi/openapi.yaml
openapi: 3.0.0
info:
  title: d4-SHA Server
  version: 1.0.0
paths:
  /tasks:
    get:
      summary: Get all tasks
      tags:
        - tasks
      responses:
        '200':
          description: A list of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
    post:
      summary: Create a new task
      tags:
        - tasks
      requestBody:
        description: The task to create
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Task'
      responses:
        '201':
          description: The created task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'

  /tasks/{id}:
    patch:
      summary: Update the task by ID
      tags:
        - tasks
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the task to update
          schema:
            type: integer
            format: uint
            minimum: 1
      requestBody:
        description: The task to update
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Task'
      responses:
        '200':
          description: The updated task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
    delete:
      summary: Delete the task by ID
      tags:
        - tasks
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the task to delete
          schema:
            type: integer
            format: uint
            minimum: 1
      responses:
          '204':
            description: The deleted task
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/Task'
  /users:
    get:
      summary: Get all users
      tags:
        - users
      responses:
        '200':
          description: A list of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
    post:
      summary: Create a new user
      tags:
        - users
      requestBody:
        description: The user to create
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '201':
          description: The created user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'

  /users/{id}:
    patch:
      summary: Update the user by ID
      tags:
        - users
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the user to delete
          schema:
            type: integer
            format: uint
            minimum: 1
      requestBody:
        description: The user to update
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: The updated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    delete:
      summary: Delete the user by ID
      tags:
        - users
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the user to delete
          schema:
            type: integer
            format: uint
            minimum: 1
      responses:
        '204':
          description: The deleted user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'

  /users/{user_id}/tasks:
    get:
      summary: Get all tasks by User ID
      operationId: GetTasksForUser
      tags:
        - users
      parameters:
        - name: user_id
          in: path
          required: true
          description: ID of the user to get tasks for
          schema:
            type: integer
            format: uint
            minimum: 1
      responses:
        '200':
          description: A list of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'

components:
  schemas:
    Task:
      type: object
      properties:
        id:
          type: integer
          format: uint # так как в БД айди это только положительные числа.
        task:
          type: string
        is_done:
          type: boolean
        user_id:
          type: integer
          format: uint
    User:
      type: object
      properties:
        id:
          type: integer
          format: uint
        email:
          type: string
        password:
          type: string